requirements.txt
Django==3.0.5
django-allauth==0.41.0
djangorestframework==3.11.0
django-cors-headers==3.2.1
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
whitenoise==5.0.1